using notifee foreground event to read initial notification and onNotificationOpenApp
